OpenAir Retrieval of Username and Project Name added.
Added retrieval of Username and Project Name based on UID and
ProjectID, respectively. In order to do so, a custom deserializer had
to be created (OpenAirDeserializer). This is because OpenAir can
possibly add custom fields server side without updating the WSDL - a
custom deserializer is necessary to suppress a SAXException when the
attribute cannot be mapped to a client side attribute in the response.